requirements.txt
Django==5.1.8
torch==2.6.0
numpy==2.2.3
djangorestframework==3.15.2
joblib==1.4.2
PyMySQL==1.1.1
dotenv==0.9.9
python-dotenv==1.1.0
PyYAML==6.0.2
scikit-learn==1.6.1
pandas==2.2.3
tqdm==4.67.1
matplotlib==3.10.1
xgboost==2.1.4
seaborn==0.13.2
scipy==1.15.2
cryptography==44.0.2
flask==3.1.0
celery[redis,msgpack]~=5.4.0
django-celery-beat==2.7.0
flower==2.0.1
djoser==2.3.1
djangorestframework_simplejwt==5.5.0
